Добавлена маршрутизация при помощи routes, добавлена панель навигации с кнопками и событиями по клику для перехода на соответствующую страницу
