+++ src/components/UI/index.js
@@ -2,10 +2,12 @@ import MyButton from "./MyButton";
 import MyInput from "./MyInput";
 import MyDialog from "./MyDialog";
 import MySelect from "./MySelect";
+import Navbar from "../Navbar";
 
 export default [
     MyButton,
     MyInput,
     MyDialog,
-    MySelect
+    MySelect,
+    Navbar
 ]
