+++ src/pages/AboutPage.vue
@@ -1,0 +1,13 @@
+<template>
+    <div><h2>Vue3 тренировочное приложение</h2></div>
+</template>
+
+<script>
+    export default {
+        name: "AboutPage"
+    }
+</script>
+
+<style scoped>
+
+</style>
