+++ src/router/router.js
@@ -1,0 +1,27 @@
+import Main from "../pages/Main";
+import {createRouter, createWebHistory} from "vue-router";
+import PostPage from "../pages/PostPage";
+import AboutPage from "../pages/AboutPage";
+
+
+const routes = [
+    {
+        path:'/',
+        component: Main
+    },
+    {
+        path: '/posts',
+        component: PostPage
+    },
+    {
+        path: '/about',
+        component: AboutPage
+    }
+]
+
+const router = createRouter({
+    routes,
+    history:createWebHistory(process.env.BASE_URL)
+})
+
+export default router;
